git-svn-id: http://www.ifountain.org/repos/os/RapidModules@1260 a08bcbf9-5f6c-4d33-af0a-e494bc2cacca

+++ RapidCMDB/grails-app/controllers/application/ApplicationController.groovy
@@ -106,8 +106,9 @@ class ApplicationController {
                 oldDomainClass.clazz.metaClass.invokeStaticMethod (oldDomainClass.clazz, "unindex", [] as Object[]);
             }
         }
-
-        FileUtils.copyDirectory (tempModelDirFile, currentModelDirFile);
+        if(tempModelDirFile.exists()){
+            FileUtils.copyDirectory (tempModelDirFile, currentModelDirFile);
+        }
         flash.message = "Reloading application."
         render(view: "application", controller: "application");
         GroovyPagesTemplateEngine.pageCache.clear();
